Add of a function in models that creates new model objects + tests.
Add of a major test based on Cosmos: tests/cosmos/distance_F/ER_1D.jl

The LHA distance F of this package is tested with Cosmos.
For several parameters in R1, R2, R3 experiments (see Bentriou,
Ballarini, Cournede 2018), Cosmos is called to
compute an average distance. Then we collect the number of simulations
and the estimated value computed by Cosmos.
We compute an average distance by the package methods with the same
number of simulations done by Cosmos

Then the two average distances are compared to each others with
regard to the width of the confidence interval set in Cosmos.
Test passed.