Forgot the notebook in the last commit.
Automaton G is implemented and runs OK. Not tested with Cosmos call.

The next step of development is the restruction of data collection for
trajectories.

From now on: Trajectory.values::Matrix{Float64}, it will be
Trajectory.values::Vector{Vector{Float64}}.

According to small performance tests I've made, it will increase the
performance of the package.