Change of Trajectory.values type from Matrix to Vector of Vector in order to replace vcat by push!.
The benchmarks are at least equal, and better when the buffer size is
small. However, no significative difference of performance with adequate
buffer size.

My opinion the gain can be seen when simulations are going to be very
long with more complicated models than we have implemented for now.

All tests passed.