A major fix in simulation was done, side effect issue, an efficient copyto! is implemented for LHA states.
Automaton G works almost well, a bias in the distance is identified
compared to Cosmos when one of the model parameters is zero.

Implementation of volatile_simulate, which simulates SynchronizedModel
without saving values: it only returns the last LHA state of the
simulation.

Implementation of distributed computations of variable automata with
volatile_simulate.

Tests passed.